napravio log out kopmonentu i funkcionlansot
